# CustomerCare/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de mon API avec Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api": {},
        "/api/showTickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get all tickets",
                "operationId": "e5ff283028dd1e15d06db1df912d0ac0",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/storeTickets": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket",
                "operationId": "a4785bf3fdf064bb8ab3e00e9c1deedf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/show/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get a specific ticket by ID",
                "operationId": "2c82c4d2541fe7631e471fc416d71a47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        },
        "/api/update/{id}": {
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update an existing ticket",
                "operationId": "38f8b681cb7e293dadc80e10c5883a61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        },
        "/api/destroy/{id}": {
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete a ticket",
                "operationId": "072645f586f0e407ce970de881ed9acb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ticket deleted successfully"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Ticket": {
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tickets",
            "description": "API Endpoints for managing tickets"
        }
    ]
}